First version of action YAML
This commit sets the first version of the bom-installer
action YAML. It is derived from the cosign installer script
with a modification to verify using the SBOM data instead
of a signature (binaries are not signed yet).

The verification code in this version is commented out as
we need at least version v0.3.0 to be able to check files
against the SBOM.

Once kubernetes artifacts get signed, we will start verifying
the SBOMs and signaures too.

Signed-off-by: Adolfo García Veytia <puerco@chainguard.dev>